26-API: Re-organize wrappers. Place interface and main impl in same dir
