+++ Okane.Api.Tests/Features/Auth/Endpoints/LoginTests.cs
@@ -17,8 +17,7 @@ using Okane.Api.Features.Auth.Endpoints;
 using Okane.Api.Features.Auth.Mappers;
 using Okane.Api.Features.Auth.Services;
 using Okane.Api.Shared.Dtos.ApiResponses;
-using Okane.Api.Shared.Wrappers.Clock;
-using Okane.Api.Shared.Wrappers.GuidGenerator;
+using Okane.Api.Shared.Wrappers;
 using Okane.Api.Tests.Testing.Integration;
 using Okane.Api.Tests.Testing.Mocks.Wrappers;
 using Okane.Api.Tests.Testing.Utils;
