+++ Okane.Api.Tests/Features/Auth/Services/TokenServiceTests.cs
@@ -9,7 +9,7 @@ using Okane.Api.Features.Auth.Entities;
 using Okane.Api.Features.Auth.Services;
 using Okane.Api.Features.Auth.Utils;
 using Okane.Api.Infrastructure.Database;
-using Okane.Api.Shared.Wrappers.Clock;
+using Okane.Api.Shared.Wrappers;
 using Okane.Api.Tests.Testing.Integration;
 using Okane.Api.Tests.Testing.Mocks.Wrappers;
 using Okane.Api.Tests.Testing.StubFactories;
