+++ Okane.Api/Shared/Wrappers/Clock.cs
@@ -1,0 +1,11 @@
+namespace Okane.Api.Shared.Wrappers;
+
+public interface IClock
+{
+    DateTime UtcNow { get; }
+}
+
+public class Clock : IClock
+{
+    public DateTime UtcNow => DateTime.UtcNow;
+}
