--- Okane.Api/Shared/Wrappers/Clock/IClock.cs
@@ -1,6 +1,0 @@
-namespace Okane.Api.Shared.Wrappers.Clock;
-
-public interface IClock
-{
-    DateTime UtcNow { get; }
-}
