--- Okane.Api/Shared/Wrappers/Clock/SystemClock.cs
@@ -1,6 +1,0 @@
-namespace Okane.Api.Shared.Wrappers.Clock;
-
-public class SystemClock : IClock
-{
-    public DateTime UtcNow => DateTime.UtcNow;
-}
